+++ modules/filesystem.py
@@ -3,8 +3,8 @@ import shutil
 import copy
 import os
 
-from helpers import relpath
-import cmd_args
+from modules.helpers import relpath
+from modules import cmd_args
 
 virtual = {}
