+++ modules/paths.py
@@ -1,6 +1,6 @@
 import os
 
-BASE_PATH = os.path.dirname(__file__)
+BASE_PATH = os.path.realpath(os.path.join(__file__, "..", ".."))
 
 def relative(*parts):
     return os.path.join(BASE_PATH, *parts)
